requirements.txt
numpy
torch
torchvision
matplotlib
paho-mqtt
flask-mqtt
pylint
autopep8